+++ app/controllers/api/frontend/excludes/groups_controller.rb
@@ -10,10 +10,14 @@ module Api
         before_action :find_excludes_group, only: %i[destroy]
 
         def create
-          case create_form.call(company: @company)
+          case create_form.call(company: @company, excludes_rules: create_excludes_rules_params)
           in { errors: errors } then render json: { errors: errors }, status: :ok
           in { result: result }
-            render json: { result: ::Excludes::GroupSerializer.new(result).serializable_hash }, status: :ok
+            render json: {
+              result: ::Excludes::GroupSerializer.new(
+                result, params: { rules: excludes_rules(result.id) }
+              ).serializable_hash
+            }, status: :ok
           end
         end
 
@@ -33,6 +37,18 @@ module Api
         def find_excludes_group
           @excludes_group = ::Excludes::Group.find_by(uuid: params[:id])
           page_not_found if @excludes_group.nil?
+        end
+
+        def excludes_rules(excludes_group_id)
+          ::Excludes::Rule
+            .where(excludes_group_id: excludes_group_id)
+            .joins(:excludes_group)
+            .hashable_pluck(:uuid, :target, :condition, :value, 'excludes_groups.uuid')
+            .group_by { |rule| rule[:excludes_groups_uuid] }
+        end
+
+        def create_excludes_rules_params
+          params.permit(excludes_rules: %i[target condition value])[:excludes_rules]
         end
       end
     end