+++ app/javascript/components/index.jsx
@@ -3,7 +3,6 @@ export * from './Company/CompanyForm';
 export * from './Company/CompanyConfiguration';
 export * from './Feedback/FeedbackForm';
 export * from './Dropdown/Dropdown';
-export * from './ExcludeRules/ExcludeRules';
 export * from './Repository/Repository';
 export * from './Repository/RepositoryForm';
 export * from './Flash/Flash';