+++ app/serializers/excludes/group_serializer.rb
@@ -3,5 +3,9 @@
 module Excludes
   class GroupSerializer < ApplicationSerializer
     set_id :uuid
+
+    attribute :excludes_rules do |object, params|
+      params[:rules][object.uuid].presence || []
+    end
   end
 end